Update User entity id int to long

+++ src/main/java/com/example/csws/config/jwt/RefreshTokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, String> {
-    Optional<RefreshToken> findByUser_Id(int userId);
+    Optional<RefreshToken> findByUser_Id(Long userId);
     Optional<RefreshToken> findByToken(String token);
-    boolean existsByUser_Id(int userId);
+    boolean existsByUser_Id(Long userId);
 }
